backend/requirements.txt
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==3.7.1
azure-ai-language-conversations==1.1.0
azure-core==1.36.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.6
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
cryptography==45.0.4
DateTime==6.0
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.115.0
Flask==3.1.2
fsspec==2026.2.0
gotrue==2.9.1
gunicorn==21.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lunardate==0.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
numpy==2.3.5
openai==1.3.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
postgrest==2.27.3
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.2
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.6.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
pypdfium2==4.30.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
pytz==2025.2
realtime==2.27.3
requests==2.32.5
rich==14.3.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.38.6
storage3==2.27.3
StrEnum==0.4.15
strictyaml==1.7.3
stripe==14.3.0
supabase==2.27.3
supabase-auth==2.27.3
supabase-functions==2.27.3
supafunc==0.6.0
tenacity==9.1.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.0
websockets==15.0.1
Werkzeug==3.1.4
yarl==1.22.0
zope.interface==8.1.1
